Listen for new activity drag and drop events in the row instead of the activity layer
